+++ strawberry/extensions/tracing/apollo.py
@@ -4,7 +4,7 @@ import dataclasses
 import time
 from datetime import datetime
 from inspect import isawaitable
-from typing import TYPE_CHECKING, Any, Dict, List, Optional
+from typing import TYPE_CHECKING, Any, Dict, Generator, List, Optional
 
 from strawberry.extensions import SchemaExtension
 from strawberry.extensions.utils import get_path_from_info
@@ -81,19 +81,19 @@ class ApolloTracingExtension(SchemaExtension):
         self._resolver_stats: List[ApolloResolverStats] = []
         self.execution_context = execution_context
 
-    def on_operation(self):
+    def on_operation(self) -> Generator[None, None, None]:
         self.start_timestamp = self.now()
         self.start_time = datetime.utcnow()
         yield
         self.end_timestamp = self.now()
         self.end_time = datetime.utcnow()
 
-    def on_parse(self):
+    def on_parse(self) -> Generator[None, None, None]:
         self._start_parsing = self.now()
         yield
         self._end_parsing = self.now()
 
-    def on_validate(self):
+    def on_validate(self) -> Generator[None, None, None]:
         self._start_validation = self.now()
         yield
         self._end_validation = self.now()
@@ -118,10 +118,10 @@ class ApolloTracingExtension(SchemaExtension):
             ),
         )
 
-    def get_results(self):
+    def get_results(self) -> Dict[str, Dict[str, Any]]:
         return {"tracing": self.stats.to_json()}
 
-    async def resolve(self, _next, root, info, *args, **kwargs):
+    async def resolve(self, _next, root, info, *args, **kwargs) -> Any:
         if should_skip_tracing(_next, info):
             result = _next(root, info, *args, **kwargs)
 
@@ -154,7 +154,7 @@ class ApolloTracingExtension(SchemaExtension):
 
 
 class ApolloTracingExtensionSync(ApolloTracingExtension):
-    def resolve(self, _next, root, info, *args, **kwargs):
+    def resolve(self, _next, root, info, *args, **kwargs) -> Any:
         if should_skip_tracing(_next, info):
             return _next(root, info, *args, **kwargs)