+++ strawberry/federation/schema.py
@@ -171,7 +171,7 @@ class Schema(BaseSchema):
         self._schema.query_type = query_type
         self._schema.type_map[query_type.name] = query_type
 
-    def entities_resolver(self, root, info, representations):
+    def entities_resolver(self, root, info, representations) -> List[object]:
         results = []
 
         for representation in representations:
@@ -341,6 +341,13 @@ class Schema(BaseSchema):
             },
             resolve=self.entities_resolver,
         )
+
+    def _warn_for_federation_directives(self) -> None:
+        # this is used in the main schema to raise if there's a directive
+        # that's for federation, but in this class we don't want to warn,
+        # since it is expected to have federation directives
+
+        pass
 
 
 def _get_entity_type(type_map: "TypeMap"):