+++ tests/asgi/conftest.py
@@ -8,31 +8,31 @@ from tests.asgi.app import create_app
 
 
 @pytest.fixture
-def test_client():
+def test_client() -> TestClient:
     app = create_app()
     return TestClient(app)
 
 
 @pytest.fixture
-def test_client_keep_alive():
+def test_client_keep_alive() -> TestClient:
     app = create_app(keep_alive=True, keep_alive_interval=0.1)
     return TestClient(app)
 
 
 @pytest.fixture
-def test_client_no_graphiql():
+def test_client_no_graphiql() -> TestClient:
     app = create_app(graphiql=False)
     return TestClient(app)
 
 
 @pytest.fixture
-def test_client_no_get():
+def test_client_no_get() -> TestClient:
     app = create_app(allow_queries_via_get=False)
     return TestClient(app)
 
 
 @pytest.fixture
-def graphql_client(test_client):
+def graphql_client(test_client) -> GraphQLTestClient:
     return GraphQLTestClient(test_client)